// src/types/Covid.ts
export interface Covid {
	message?: any;

	updated: string;
	country?: string;
	countryInfo?: CoutryInfo;
	cases: number;
	todayCases: number;
	deaths: number;
	todayDeaths: number;
	recovered: number;
	todayRecovered: number;
	active: number;
	critical: number;
	casesPerOneMillion: number;
	deathsPerOneMillion: number;
	tests: number;
	testsPerOneMillion: number;
	population?: number;
	continent?: string;
	oneCasePerPeople: number;
	oneDeathPerPeople: number;
	oneTestPerPeople: number;
	activePerOneMillion: number;
	recoveredPerOneMillion: number;
	criticalPerOneMillion: number;
}

interface CoutryInfo {
	_id: number;
	iso2: string;
	iso3: string;
	lat: number;
	long: number;
	flag: string;
}